性能大幅度提升，2w假人实测性能正常，体验流程稳定

+++ src/M2Server/Map/Envir.cs
@@ -7,7 +7,7 @@ namespace M2Server
     {
         public byte btType;
         public object CellObj;
-        public double dwAddTime;
+        public int dwAddTime;
         public bool boObjectDisPose;
     }
